Added bat, corrected SBA with skrink one correct for the PFC sutdy fixed for rat study

+++ fonctions/_9_coregistration_to_template_space.py
@@ -1,8 +1,17 @@
 import os
 import subprocess
-import glob
 from fonctions.extract_filename import extract_filename
 import ants
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKCYAN = '\033[96m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
 
 #Path to the excels files and data structure
 opj = os.path.join
@@ -12,7 +21,6 @@ opd = os.path.dirname
 ope = os.path.exists
 spco = subprocess.check_output
 spgo = subprocess.getoutput
-
 
 ####################################################################################
 ########################## Step 3 normalisation to template atlas space ############
@@ -40,16 +48,15 @@ def to_common_template_space(deoblique, dir_fMRI_Refth_RS_prepro1, dir_fMRI_Reft
         elif do_anat_to_func == False and anat_func_same_space == True:
             mvt_shft_ANTs = []
             w2inv_fwd = []
-        else : print('ERROR: If Anat and Func are not in the same space you need to perform that trasnformation (do_anat_to_func = True)')
+        else :
+            raise Exception(bcolors.FAIL + 'ERROR: If Anat and Func are not in the same space you need to perform that trasnformation (do_anat_to_func = True)' + bcolors.ENDC)
 
     ##### creat new variable  for template space
     if ope(dir_fMRI_Refth_RS_prepro3) == False:
         os.makedirs(dir_fMRI_Refth_RS_prepro3)
+    if opj(dir_fMRI_Refth_RS_prepro3,'tmp') == False:
         os.makedirs(opj(dir_fMRI_Refth_RS_prepro3,'tmp'))
 
-        ###############################################################################################
-        ################################### registration to anat space ################################
-        ###############################################################################################
         ############################### ############################### ###############################
         ############################### apply transfo to anat space to Mean_Image image for test ######
         ############################### ############################### ###############################
@@ -62,39 +69,40 @@ def to_common_template_space(deoblique, dir_fMRI_Refth_RS_prepro1, dir_fMRI_Reft
         ' -prefix ' + output2 + ' -expr "a"'
         spco([command], shell=True)
 
+        r = REF_int
+        root_RS = extract_filename(RS[r])
+        command = 'export SINGULARITYENV_AFNI_NIFTI_TYPE_WARN="NO";singularity run' + s_bind + afni_sif + '3dinfo -orient ' + opj(dir_fMRI_Refth_RS_prepro1, root_RS + '_xdtrf_mean_preWARP' + '.nii.gz') + ' ' +  output2
+        orientation_orig = spgo(command).split('\n')[-1]
+        print(bcolors.OKGREEN + 'orientation orig of the func image is ' + orientation_orig + bcolors.ENDC)
+        ##### apply the recenter fmri
+        #######!!!!!! I don't know why if deobblique1=WARP no gridset if exeption2 require gridset!!!!!
+        if deoblique == 'header':
+            command = 'singularity run' + s_bind + afni_sif + '3drefit -deoblique ' + overwrite + ' -orient ' + orientation + ' ' + output2
+            spco([command], shell=True)
+
+        elif deoblique == 'WARP' or deoblique == 'WARP_without_3drefit':
+            # reorient the fiedls according to the json file
+            command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + \
+            ' ' +  output2
+            spco([command], shell=True)
+
+        elif deoblique == 'WARP_Gridset':  # do nothing
+            # reorient the fiedls according to the json file
+            command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + ' -gridset ' + output2 + \
+                      ' ' + output2
+            spco([command], shell=True)
+
+        elif deoblique == 'header_WO_deob':
+            command = 'singularity run' + s_bind + afni_sif + '3drefit ' + overwrite + ' -orient ' + orientation + ' ' + output2
+            spco([command], shell=True)
+
+        elif deoblique == 'no_deoblique':  # do nothing
+            print('')
+
+        else:
+            raise Exception(bcolors.FAIL + 'ERROR: wrong deoblique name' + bcolors.ENDC)
+
         if anat_func_same_space == True:
-            r = REF_int
-            root_RS = extract_filename(RS[r])
-            command = 'export SINGULARITYENV_AFNI_NIFTI_TYPE_WARN="NO";singularity run' + s_bind + afni_sif + '3dinfo -orient ' + opj(dir_fMRI_Refth_RS_prepro1, root_RS + '_xdtrf_mean_preWARP' + '.nii.gz') + ' ' +  output2
-            orientation_orig = spgo(command).split('\n')[-1]
-            print(orientation_orig)
-            ##### apply the recenter fmri
-            #######!!!!!! I don't know why if deobblique1=WARP no gridset if exeption2 require gridset!!!!!
-            if deoblique == 'header':
-                command = 'singularity run' + s_bind + afni_sif + '3drefit -deoblique ' + overwrite + ' -orient ' + orientation + ' ' + output2
-                spco([command], shell=True)
-
-            elif deoblique == 'WARP' or deoblique == 'WARP_without_3drefit':
-                # reorient the fiedls according to the json file
-                command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + \
-                ' ' +  output2
-                spco([command], shell=True)
-
-            elif deoblique == 'WARP_Gridset':  # do nothing
-                print('WARP_Gridset')
-                # reorient the fiedls according to the json file
-                command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + ' -gridset ' + output2 + \
-                          ' ' + output2
-                spco([command], shell=True)
-
-            elif deoblique == 'header_WO_deob':
-                print('header_WO_deob')
-                command = 'singularity run' + s_bind + afni_sif + '3drefit ' + overwrite + ' -orient ' + orientation + ' ' + output2
-                spco([command], shell=True)
-
-            elif deoblique == 'no_deoblique':  # do nothing
-                print('no_deoblique')
-
             command = 'singularity run' + s_bind + afni_sif + '3dZeropad -I 50 -S 50 -A 50 -P 50 -L 50 -R 50 -S 50 -prefix ' +  output2 + ' ' +  output2 + ' -overwrite'
             spco([command], shell=True)
 
@@ -103,35 +111,17 @@ def to_common_template_space(deoblique, dir_fMRI_Refth_RS_prepro1, dir_fMRI_Reft
             ' -input  ' + output2
             #' -master ' + opj(dir_prepro, ID + '_acpc_cropped' + TfMRI + '.nii.gz')
             spco([command], shell=True)
-            '''
-            command = '3dinfo -di ' + input2
-            delta_x = str(abs(round(float(spgo([command])[-8:]), 10)))
-            command = '3dinfo -dj ' + input2
-            delta_y = str(abs(round(float(spgo([command])[-8:]), 10)))
-            command = '3dinfo -dk ' + input2
-            delta_z = str(abs(round(float(spgo([command])[-8:]), 10)))
-
-            command = '3dresample' + overwrite + \
-                      ' -prefix ' + output2 + \
-                      ' -dxyz ' + delta_x + ' ' + delta_y + ' ' + delta_z + ' ' + \
-                      ' -input  ' + output2
-            spco([command], shell=True)
-        '''
-        else:
-            print('nothing to do')
 
         ## test on mean img (to see spatially that is works)
-        print("starting est on mean image and anat test to standart space")
+        print(bcolors.OKGREEN + "starting func to template space on mean image and anat test transormation" + bcolors.ENDC)
         MEAN = ants.image_read(output2)
         REF  = ants.image_read(opj(dir_fMRI_Refth_RS_prepro3,'BASE_SS_fMRI.nii.gz'))
-        print(mvt_shft_ANTs + transfo_concat_Anat)
-        print(w2inv_Anat + w2inv_fwd)
         TRANS = ants.apply_transforms(fixed=REF, moving=MEAN,
                                       transformlist=transfo_concat_Anat + mvt_shft_ANTs,
                                       interpolator='nearestNeighbor',
                                       whichtoinvert=w2inv_Anat + w2inv_fwd)
         ants.image_write(TRANS, output3, ri=False)
-        print(str(output3) + ' done!')
+        print(bcolors.OKGREEN + str(output3) + ' done!' + bcolors.ENDC)
 
             ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ##
             ## ## ## ## ## ## ## ## ## ## ## ## ## ##  Work on all FUNC ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ## ##
@@ -139,9 +129,10 @@ def to_common_template_space(deoblique, dir_fMRI_Refth_RS_prepro1, dir_fMRI_Reft
     for i in range(0, int(nb_run)):
         ##### go for BOLD img preTTT
         root_RS = extract_filename(RS[i])
-        for f in os.listdir(opj(dir_fMRI_Refth_RS_prepro3, 'tmp')):
-            if ope(os.path.join(opj(dir_fMRI_Refth_RS_prepro3, 'tmp'), f)):
-                os.remove(os.path.join(opj(dir_fMRI_Refth_RS_prepro3, 'tmp'), f))
+        if ope(opj(dir_fMRI_Refth_RS_prepro3, 'tmp')):
+            for f in os.listdir(opj(dir_fMRI_Refth_RS_prepro3, 'tmp')):
+                if ope(os.path.join(opj(dir_fMRI_Refth_RS_prepro3, 'tmp'), f)):
+                    os.remove(os.path.join(opj(dir_fMRI_Refth_RS_prepro3, 'tmp'), f))
 
         output3 = opj(dir_fMRI_Refth_RS_prepro3, root_RS + '_residual_in_anat_sfht.nii.gz')
         output2 = opj(dir_fMRI_Refth_RS_prepro3, root_RS + '_residual_in_anat_reorient.nii.gz')
@@ -151,14 +142,15 @@ def to_common_template_space(deoblique, dir_fMRI_Refth_RS_prepro1, dir_fMRI_Reft
             output3 = opj(dir_fMRI_Refth_RS_prepro3, root_RS + '_3dDeconvolve_failed_in_anat_sfht.nii.gz')
             output2 = opj(dir_fMRI_Refth_RS_prepro3, root_RS + '_3dDeconvolve_failed_in_anat_reorient.nii.gz')
             input2 = opj(dir_fMRI_Refth_RS_prepro1, root_RS + '_3dDeconvolve_failed.nii.gz')
-        else:
-            print('ERROR: corrected func image in func space not found')
+            if ope(input2) == False:
+                    raise Exception(bcolors.FAIL + 'ERROR: corrected func image in func space not found' + bcolors.ENDC)
+
+        command = 'singularity run' + s_bind + afni_sif + '3dcalc' + overwrite + ' -a ' + input2 + \
+                  ' -prefix ' + output2 + ' -expr "a"'
+        spco([command], shell=True)
+
 
         if anat_func_same_space == True:
-            command = 'singularity run' + s_bind + afni_sif + '3dcalc' + overwrite + ' -a ' + input2 + \
-                      ' -prefix ' + output2 + ' -expr "a"'
-            spco([command], shell=True)
-
             command = 'export SINGULARITYENV_AFNI_NIFTI_TYPE_WARN="NO";singularity run' + s_bind + afni_sif + '3dinfo -same_obl ' + opj(dir_prepro, ID + 'template_indiv' + TfMRI + '.nii.gz') + ' ' +  output2
             nx = spgo(command).split('\n')[-1]
 
@@ -166,35 +158,35 @@ def to_common_template_space(deoblique, dir_fMRI_Refth_RS_prepro1, dir_fMRI_Reft
             root_RS = extract_filename(RS[r])
             command = 'export SINGULARITYENV_AFNI_NIFTI_TYPE_WARN="NO";singularity run' + s_bind + afni_sif + '3dinfo -orient ' + opj(dir_fMRI_Refth_RS_prepro1, root_RS + '_xdtrf_mean_preWARP' + '.nii.gz') + ' ' +  output2
             orientation_orig = spgo(command).split('\n')[-1]
-            print(orientation_orig)
 
-            ##### apply the recenter fmri
-            #######!!!!!! I don't know why if deobblique1=WARP no gridset if exeption2 require gridset!!!!!
-            if deoblique == 'header':
-                command = 'singularity run' + s_bind + afni_sif + '3drefit -deoblique ' + overwrite + ' -orient ' + orientation + ' ' + output2
-                spco([command], shell=True)
+        ##### apply the recenter fmri
+        #######!!!!!! I don't know why if deobblique1=WARP no gridset if exeption2 require gridset!!!!!
+        if deoblique == 'header':
+            command = 'singularity run' + s_bind + afni_sif + '3drefit -deoblique ' + overwrite + ' -orient ' + orientation + ' ' + output2
+            spco([command], shell=True)
 
-            elif deoblique == 'WARP' or deoblique == 'WARP_without_3drefit':
-                # reorient the fiedls according to the json file
-                command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + \
-                          ' ' + output2
-                spco([command], shell=True)
+        elif deoblique == 'WARP' or deoblique == 'WARP_without_3drefit':
+            # reorient the fiedls according to the json file
+            command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + \
+                      ' ' + output2
+            spco([command], shell=True)
 
-            elif deoblique == 'WARP_Gridset':  # do nothing
-                print('WARP_Gridset')
-                # reorient the fiedls according to the json file
-                command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + ' -gridset ' + output2 + \
-                          ' ' + output2
-                spco([command], shell=True)
+        elif deoblique == 'WARP_Gridset':  # do nothing
+            # reorient the fiedls according to the json file
+            command = 'singularity run' + s_bind + afni_sif + '3dWarp' + overwrite + ' -deoblique -NN -prefix ' + output2 + ' -gridset ' + output2 + \
+                      ' ' + output2
+            spco([command], shell=True)
 
-            elif deoblique == 'header_WO_deob':
-                print('header_WO_deob')
-                command = 'singularity run' + s_bind + afni_sif + '3drefit ' + overwrite + ' -orient ' + orientation + ' ' + output2
-                spco([command], shell=True)
+        elif deoblique == 'header_WO_deob':
+            command = 'singularity run' + s_bind + afni_sif + '3drefit ' + overwrite + ' -orient ' + orientation + ' ' + output2
+            spco([command], shell=True)
 
-            elif deoblique == 'no_deoblique':  # do nothing
-                print('no_deoblique')
+        elif deoblique == 'no_deoblique':  # do nothing
+            print('')
+        else:
+            raise Exception(bcolors.FAIL + 'ERROR: wrong deoblique name' + bcolors.ENDC)
 
+        if anat_func_same_space == True:
             command = 'singularity run' + s_bind + afni_sif + '3dZeropad -I 50 -S 50 -A 50 -P 50 -L 50 -R 50 -S 50 -prefix ' + output3 + ' ' + output2 + ' -overwrite'
             spco([command], shell=True)
 
@@ -247,7 +239,7 @@ def to_common_template_space(deoblique, dir_fMRI_Refth_RS_prepro1, dir_fMRI_Reft
                   ' -input  ' + BASE_SS_mask
         spco([command], shell=True)
     else:
-        print('WARNING: list_atlases is empty!')
+        print(bcolors.WARNING + 'WARNING: list_atlases is empty!' + bcolors.ENDC)
 
     if creat_study_template== True:
         command = 'singularity run' + s_bind + afni_sif + '3dresample' + overwrite + \
